+++ src/lib/abstraction/form-control.ts
@@ -1,20 +1,12 @@
 import * as ng from '@angular/forms';
+import {InitHandler} from './init-handler';
 
 export class FormControl extends ng.FormControl {
 
-  get isInitialized(): boolean {
-    return this._isInitialized;
-  }
-
-  set isInitialized(value: boolean) {
-    this._isInitialized = value;
-  }
-
-  // tslint:disable-next-line:variable-name
-  private _isInitialized: boolean = false;
+  public readonly initHandler: InitHandler<FormControl> = new InitHandler<FormControl>(this);
 
   public updateValueAndValidity(opts?: { onlySelf?: boolean; emitEvent?: boolean }): void {
-    if (this._isInitialized) {
+    if (!this.initHandler || this.initHandler.isInitialized) {
       super.updateValueAndValidity(opts);
     }
   }