+++ src/lib/abstraction/form-group.ts
@@ -1,24 +1,12 @@
 import * as ng from '@angular/forms';
+import {InitHandler} from './init-handler';
 
 export class FormGroup extends ng.FormGroup {
-  get isInitialized(): boolean {
-    return this._isInitialized;
-  }
 
-  set isInitialized(value: boolean) {
-    const ripple = this._isInitialized !== value;
-    this._isInitialized = value;
-
-    if (ripple) {
-      Object.keys(this.controls).forEach((k) => (this.controls[k] as any).isInitialized = value);
-    }
-  }
-
-  // tslint:disable-next-line:variable-name
-  private _isInitialized: boolean = false;
+  public readonly initHandler: InitHandler<FormGroup> = new InitHandler<FormGroup>(this);
 
   public updateValueAndValidity(opts?: { onlySelf?: boolean; emitEvent?: boolean }): void {
-    if (this._isInitialized) {
+    if (!this.initHandler || this.initHandler.isInitialized) {
       super.updateValueAndValidity(opts);
     }
   }