+++ src/lib/form-group/form-group.component.ts
@@ -9,11 +9,11 @@ import {
   QueryList,
   SimpleChanges
 } from '@angular/core';
-import {FormModel} from '../form-builder/form-builder.component';
 import {FormSlotComponent} from '../form-slot/form-slot.component';
 import {FormFieldComponent} from '../form-field/form-field.component';
 import {AbstractFormFieldComponent} from '../abstract-form-field/abstract-form-field.component';
 import {FormArray, FormControl, FormGroup} from '../abstraction';
+import {FormModel} from '../form-builder/form-builder.component';
 
 @Component({
   selector: 'mw-form-group',
@@ -31,12 +31,24 @@ import {FormArray, FormControl, FormGroup} from '../abstraction';
     </ng-container>
   `,
 })
-export class FormGroupComponent extends AbstractFormFieldComponent<any> implements OnChanges, AfterViewInit {
+export class FormGroupComponent
+  extends AbstractFormFieldComponent<any>
+  implements OnChanges, AfterViewInit {
+
+  @Input()
+  public isRootGroup: boolean = false;
+
   @Input()
   public model: FormModel;
 
   @ContentChildren(FormSlotComponent, {descendants: true})
   public mySlots: QueryList<FormSlotComponent>;
+
+  public renderTargets: { [key: string]: FormSlotComponent } = {};
+
+  private initializeTimeoutHandle: number = null;
+
+  public orderAsGiven = (): number => 1;
 
   public get groupPath(): string {
     return (this.path == null || this.path.trim() === '')
@@ -52,17 +64,13 @@ export class FormGroupComponent extends AbstractFormFieldComponent<any> implemen
     let p = this.element;
     do {
       p = p.parent;
-      if (p != null && p instanceof FormArray) {
+      if (p instanceof FormArray) {
         return this.index;
       }
-    } while (p != null && p.parent != null && !(parent instanceof FormControl));
+    } while (p && p.parent && !(parent instanceof FormControl));
 
     return null;
   }
-
-  public renderTargets: { [key: string]: FormSlotComponent } = {};
-
-  public orderAsGiven = (): number => 1;
 
   constructor(private readonly cfr: ComponentFactoryResolver,
               private readonly cdr: ChangeDetectorRef,
@@ -71,29 +79,29 @@ export class FormGroupComponent extends AbstractFormFieldComponent<any> implemen
   }
 
   public ngAfterViewInit(): void {
-    if (this.mySlots != null) {
+    if (this.mySlots) {
       this.renderElementsInSlots();
+      this.cdr.detectChanges();
     }
-    this.cdr.detectChanges();
   }
 
   public ngOnChanges(changes: SimpleChanges): void {
     let wasRendered = false;
-    if (changes.hasOwnProperty('slots') && changes.slots.currentValue != null) {
+    if ('slots' in changes && changes.slots.currentValue) {
       this.renderElementsInSlots();
       wasRendered = true;
     }
-    if (!wasRendered && changes.hasOwnProperty('mySlots') && changes.mySlots.currentValue != null) {
+    if (!wasRendered && 'mySlots' in changes && changes.mySlots.currentValue) {
       this.renderElementsInSlots();
     }
   }
 
   private renderElementsInSlots(): void {
-    if (this.slots != null || this.mySlots != null) {
+    if (this.slots || this.mySlots) {
       this.createRenderTargets();
 
       Object.keys(this.renderTargets).map((name) => {
-        if (!this.renderTargets.hasOwnProperty(name)) {
+        if (!(name in this.renderTargets)) {
           return;
         }
         this.renderTargets[name].viewRef.clear();
@@ -109,14 +117,23 @@ export class FormGroupComponent extends AbstractFormFieldComponent<any> implemen
         target.instance.index = this.indexFromParent;
 
         target.instance.render();
-        (this.element as FormGroup).isInitialized = true;
       });
+
+      if (this.isRootGroup) {
+        if (this.initializeTimeoutHandle) {
+          window.clearTimeout(this.initializeTimeoutHandle);
+        }
+        this.initializeTimeoutHandle = window.setTimeout(() => {
+          (this.element as FormGroup).initHandler.setIsInitialized(true);
+          this.cdr.detectChanges();
+        }, 5);
+      }
     }
   }
 
   private createRenderTargets(): void {
     const groupPath = this.groupPath;
-    if (this.slots != null) {
+    if (this.slots) {
       this.slots.toArray().forEach((slot) => {
         if (slot.fieldName.substr(0, groupPath.length) !== groupPath) {
           return;
@@ -131,7 +148,7 @@ export class FormGroupComponent extends AbstractFormFieldComponent<any> implemen
       });
     }
 
-    if (this.mySlots != null) {
+    if (this.mySlots) {
       this.mySlots.toArray().forEach((slot) => {
         this.renderTargets[slot.fieldName] = slot;
       });